requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pafy == 0.5.5
python_vlc == 3.0.12118
rich == 10.3.0
youtube_search_python == 1.4.5
inquirer